Do not eager load at the model level
It's not great to eager load stuff at the model level. In case a
previous user of the data has preloaded already, we're now preloading
twice. Since preloading is expensive, we should avoid that and commit to
only loading from model or service objects when there's no controller we
can hook into.