Fix PictureTransformations#inferred_sizes_from_string
We want to infer the correct height or width given that either height or
width are not given. In this case we (correctly) calculate the original
aspect ratio of the image, and then infer the width or height
respectively from the given width/height and the ration.

The previous implementation could have been more simple:

```
if sizes[:height].zero?
  sizes[:width] = image_file_width
else
  sizes[:height] = image_file_height
end
```

But that implementation does not make that much sense.

Additionally I'm rounding things here, because DragonFly does not like
decimal points in their transformation strings.